Fix max mpi and max omp setting for non-default folders
The setting of maximum number of MPI ranks and OpenMP threads did
not work when the tests were not run from the default location, the
root directory of the regression tests. This change fixes this by
always looking for the respective file in the input directory rather
than the working directory.